Rename AvaloniaXamlIlRelativePointAstNode to generalized AvaloniaXamlIlVisualElementAstNode
